# src/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Users segments",
        "title": "Users segments service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/reports/{reportID}": {
            "get": {
                "description": "Download a report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Download",
                "operationId": "download",
                "parameters": [
                    {
                        "type": "string",
                        "example": "80ef1ba7-1045-41aa-a8a2-4c0aba407baf",
                        "description": "user ID",
                        "name": "reportID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/segments": {
            "get": {
                "description": "Returns all segments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "Get",
                "operationId": "getSegments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.segmentsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new segment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "Create",
                "operationId": "create",
                "parameters": [
                    {
                        "description": "Segment data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.createSegmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.createSegmentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/segments/:segmentSlug": {
            "delete": {
                "description": "Deletes segment and all users from it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "Delete",
                "operationId": "delete",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1",
                        "description": "segment ID",
                        "name": "segmentSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register",
                "operationId": "register",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.registerUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.registerUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/users/:userID/segments": {
            "get": {
                "description": "Returns all users segments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get users segments",
                "operationId": "getUsersSegments",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1",
                        "description": "user ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.usersSegmentsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/users/:userID/segments/add": {
            "post": {
                "description": "Adds user to segments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Add to segments",
                "operationId": "add",
                "parameters": [
                    {
                        "description": "Segments to add",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.modifyUsersSegmentsRequest"
                        }
                    },
                    {
                        "type": "string",
                        "example": "1",
                        "description": "user ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.usersSegmentsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/users/:userID/segments/delete": {
            "post": {
                "description": "Deletes user from segment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete from segment",
                "operationId": "deleteFromSegment",
                "parameters": [
                    {
                        "description": "Segments to delete",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.modifyUsersSegmentsRequest"
                        }
                    },
                    {
                        "type": "string",
                        "example": "1",
                        "description": "user ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.usersSegmentsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/users/:userID/segments/report": {
            "get": {
                "description": "Returns link to a csv with user segments report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get users segments report",
                "operationId": "getUsersSegmentsReport",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1",
                        "description": "user ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "8",
                        "description": "month",
                        "name": "month",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2023",
                        "description": "year",
                        "name": "year",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.ReportResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "v1.ReportResponse": {
            "type": "object",
            "properties": {
                "fileLink": {
                    "type": "string",
                    "example": "http://localhost:8080/v1/reports/80ef1ba7-1045-41aa-a8a2-4c0aba407baf"
                }
            }
        },
        "v1.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "v1.createSegmentRequest": {
            "type": "object",
            "required": [
                "slug"
            ],
            "properties": {
                "slug": {
                    "type": "string",
                    "example": "AVITO_VOICE_MESSAGES"
                }
            }
        },
        "v1.createSegmentResponse": {
            "type": "object",
            "required": [
                "id",
                "slug"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "slug": {
                    "type": "string",
                    "example": "AVITO_VOICE_MESSAGES"
                }
            }
        },
        "v1.modifyUsersSegmentsRequest": {
            "type": "object",
            "required": [
                "segments"
            ],
            "properties": {
                "segments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "AVITO_VOICE_MESSAGES",
                        "AVITO_PERFORMANCE_VAS"
                    ]
                }
            }
        },
        "v1.registerUserRequest": {
            "type": "object",
            "required": [
                "email",
                "lastName",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "vadiminmail@gmail.com"
                },
                "lastName": {
                    "type": "string",
                    "example": "Valov"
                },
                "name": {
                    "type": "string",
                    "example": "Vadim"
                }
            }
        },
        "v1.registerUserResponse": {
            "type": "object",
            "required": [
                "email",
                "id",
                "lastName",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "vadiminmail@gmail.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "lastName": {
                    "type": "string",
                    "example": "Valov"
                },
                "name": {
                    "type": "string",
                    "example": "Vadim"
                }
            }
        },
        "v1.segmentsResponse": {
            "type": "object",
            "properties": {
                "segments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "AVITO_VOICE_MESSAGES"
                    ]
                }
            }
        },
        "v1.usersSegmentsResponse": {
            "type": "object",
            "properties": {
                "usersSegments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "AVITO_VOICE_MESSAGES",
                        "AVITO_PERFORMANCE_VAS"
                    ]
                }
            }
        }
    }
}